refactor minor
